% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reproducer.R
\name{boxplotAndDensityCurveOnHistogram}
\alias{boxplotAndDensityCurveOnHistogram}
\title{boxplotAndDensityCurveOnHistogram}
\usage{
boxplotAndDensityCurveOnHistogram(df, colName, limLow, limHigh)
}
\arguments{
\item{df}{Data frame with data to be displayed}

\item{colName}{Name of the selected column in a given data frame}

\item{limLow}{the limit on the lower side of the displayed range}

\item{limHigh}{the limit on the higher side of the displayed range}
}
\value{
A figure being a density curve overlaid on histogram
}
\description{
Boxplot and density curve overlaid on histogram
}
\examples{
library(grid)
library(gridExtra)
library(ggplot2)
boxplotAndDensityCurveOnHistogram(Madeyski15EISEJ.PropProjects, "STUD", 0, 100)
boxplotAndDensityCurveOnHistogram(Madeyski15SQJ.NDC, "simple", 0, 100)
}
\author{
Lech Madeyski
}

