% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducer.R
\name{boxplotHV}
\alias{boxplotHV}
\title{boxplotHV}
\usage{
boxplotHV(df, colName, limLow, limHigh, isHorizontal)
}
\arguments{
\item{df}{Data frame with data to be displayed}

\item{colName}{Name of the selected column in a given data frame}

\item{limLow}{the limit on the lower side of the displayed range}

\item{limHigh}{the limit on the higher side of the displayed range}

\item{isHorizontal}{Boolean value to control whether the box plot should be horizontal or not (i.e., vertical)}
}
\value{
A box plot
}
\description{
Box plot
}
\examples{
boxplotHV(Madeyski15EISEJ.PropProjects, "STUD", 0, 100, TRUE)
boxplotHV(Madeyski15EISEJ.PropProjects, "STUD", 0, 100, FALSE)
boxplotHV(Madeyski15SQJ.NDC, "simple", 0, 100, FALSE)
boxplotHV(Madeyski15SQJ.NDC, "simple", 0, 100, TRUE)
}
\author{
Lech Madeyski
}
