% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPSimulation.R
\name{calc.a}
\alias{calc.a}
\title{calc.a}
\usage{
calc.a(f, A)
}
\arguments{
\item{f}{a vector defining the degrees of freedom for the effect sizes}

\item{A}{a vector defining the constants that relate each StdMD to its related t-variable where t^2=Ad^2}
}
\value{
The value of a.
}
\description{
This function is a helper function that calculates one element of the standardized mean difference effect size variance based on Hedges and Olkin p128-131.
}
\examples{
reproducer:::calc.a(10,2/10)
# [1] 0.2128649
}
\author{
Barbara Kitchenham and Lech Madeyski
}
