% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPSimulation.R
\name{calc.b}
\alias{calc.b}
\title{calc.b}
\usage{
calc.b(f)
}
\arguments{
\item{f}{a vector defining the degrees of freedom for the effect sizes}
}
\value{
The value of b.
}
\description{
This function is a helper function that calculates one element of the standardized mean difference effect size variance based on Hedges and Olkin p128-131.
}
\examples{
reproducer:::calc.b(8)
#0.08649774

}
\author{
Barbara Kitchenham and Lech Madeyski
}
