% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPSimulation.R
\name{CatchError}
\alias{CatchError}
\title{CatchError}
\usage{
CatchError(expr)
}
\arguments{
\item{expr}{The expression that is being monitored}
}
\value{
A message confirming whether the expression has performed
successfully
}
\description{
This is a helper function to stop simulations failing if the
metafor function rma fails for example cannot converge properly for
a specific dataset.
}
\examples{
ES=c(0.2,0.3)
ESvar=c(0.04,0.03)
outcome=reproducer:::CatchError(metafor::rma(ES,ESvar,method='Meth'))
outcome
# [1] 'Failure'
}
\author{
Barbara Kitchenham and Lech Madeyski
}
