% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPSimulation.R
\name{checkIfValidDummyVariable}
\alias{checkIfValidDummyVariable}
\title{checkIfValidDummyVariable}
\usage{
checkIfValidDummyVariable(vector)
}
\arguments{
\item{vector}{Vector variable}
}
\value{
The logical value:
TRUE - if a vector variable passed as the function's parameter represents a
valid dummy variable, i.e., comprises only zeros and 1's.
FALSE - if a vector variable passed as the function's parameter does not
represent a valid dummy variable.
}
\description{
This helper function checks whether a vector variable comprises
only zeros and 1's.
}
\examples{
print(reproducer:::checkIfValidDummyVariable(c(0, 1, 0, 0)))
# [1] TRUE
print(reproducer:::checkIfValidDummyVariable(c(0, 1, 2, 0)))
# [1] FALSE
}
\author{
Barbara Kitchenham and Lech Madeyski
}
