% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reproducer.R
\name{cloudOfWords}
\alias{cloudOfWords}
\title{cloudOfWords}
\usage{
cloudOfWords(textFile)
}
\arguments{
\item{textFile}{A text file used to produce a word cloud}
}
\value{
A figure being a word cloud
}
\description{
cloud of words
}
\examples{
myPath=system.file("NAMESPACE", package = "reproducer")
cloudOfWords(textFile=myPath)
}
\author{
Lech Madeyski
}

