% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSendQueries.R
\name{dbSendQueries}
\alias{dbSendQueries}
\title{Function to send sql queries to the provenance database}
\usage{
dbSendQueries(con, sql)
}
\arguments{
\item{con}{Connection to the provenance database}

\item{sql}{SQL query}
}
\description{
This contains the function to send sql queries to create
the provenance database on the model decribed on db_schema.sql.
}
\keyword{creation}
\keyword{database,}
\keyword{sql,}
