% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducer.R
\name{densityCurveOnHistogram}
\alias{densityCurveOnHistogram}
\title{densityCurveOnHistogram}
\usage{
densityCurveOnHistogram(df, colName, limLow, limHigh)
}
\arguments{
\item{df}{Data frame with data to be displayed}

\item{colName}{Name of the selected column in a given data frame}

\item{limLow}{the limit on the lower side of the displayed range}

\item{limHigh}{the limit on the higher side of the displayed range}
}
\value{
A figure being a density curve overlaid on histogram
}
\description{
Density curve overlaid on histogram
}
\examples{
densityCurveOnHistogram(Madeyski15EISEJ.PropProjects, "STUD", 0, 100)
# densityCurveOnHistogram(data.frame(x<-rnorm(50, mean=50, sd=5)), 'x', 0, 100)
}
\author{
Lech Madeyski
}
