% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPSimulation.R
\name{doLM}
\alias{doLM}
\title{doLM}
\usage{
doLM(DataSet, Metric, Type)
}
\arguments{
\item{DataSet}{The dataset to be analysed.}

\item{Metric}{The name of the variable to be analysed}

\item{Type}{Defines the experimental design}
}
\value{
The data analysis results provided by the lmer function of the lme4 package
}
\description{
This helper function is called by the function 'crossoverResidualAnalysis' to perform either an AB/BA crossover analysis or a four-group crossover on the data set defined by the parameter DataSet depending on the value of the Type parameter.
}
\author{
Barbara Kitchenham and Lech Madeyski
}
