% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPaperData.R
\docType{data}
\name{KitchenhamEtAl.CorrelationsAmongParticipants.Gravino15JVLC}
\alias{KitchenhamEtAl.CorrelationsAmongParticipants.Gravino15JVLC}
\title{KitchenhamEtAl.CorrelationsAmongParticipants.Gravino15JVLC data illustrate correlations between results
from individual participants in a family of 2 cross-over experiments conducted by Gravino et al.:
[1] C. Gravino, G. Scanniello, and G. Tortora, "Source-code comprehension tasks supported by UML design models:
Results from a controlled experiment and a differentiated replication,"
Journal of Visual Languages and Computing, vol. 28, pp. 23–38, 2015.
The experiments assess whether the comprehension of object oriented source-code increases used
with UML class and sequence diagrams produced in the software design phase.
If you use this data set please cite:
[1] C. Gravino, G. Scanniello, and G. Tortora, "Source-code comprehension tasks supported by UML
design models: Results from a controlled experiment and a differentiated replication,"
Journal of Visual Languages and Computing, vol. 28, pp. 23–38, 2015.
[2] Barbara Kitchenham, Lech Madeyski, Giuseppe Scanniello and Carmine Gravino, "The importance
of the Correlation between Results from Individual Participants in Crossover Experiments"
(to be submitted as of 2020).}
\format{
A data frame with 64  rows and 9 variables:
\describe{
\item{ExperimentID}{<fct>|ExperimentID: A unique identifier for each of the three experiments
in the data set.}
\item{ParticipantID}{<fct>|Participant ID: An identifier for each participant, unique for a
specific experiment.}
\item{SequenceGroup}{<fct>|Experimental Sequence Group: A , B , C, D}
\item{System}{<fct>|Software systems used in the experiment: Music shop, a system for handling
the sales of a music shop. Theater ticket, a system for managing theatre reservations.  }
\item{Period}{<fct>|Time period of the cross-over experiment: 1 or 2}
\item{Treatment}{<fct>|Experimental Treatment: Mo, design models were available, NOMo design
models were not available}
\item{Comprehension}{<dbl>|Dependent variable: The level of comprehension achieved by the
software engineer. }
\item{Time }{<dbl>|Dependent variable: The time [min] taken to complete the comprehension task. }
\item{CrossOverID}{<fct>|CrossOver category: For 4 group crossover designs, the crossover
category specifies the matching pairs of sequence groups, CO1 and CO2. For 2 group crossover,
the category is set to CO1 only}
}
}
\usage{
KitchenhamEtAl.CorrelationsAmongParticipants.Gravino15JVLC
}
\description{
KitchenhamEtAl.CorrelationsAmongParticipants.Gravino15JVLC data illustrate correlations between results
from individual participants in a family of 2 cross-over experiments conducted by Gravino et al.:
[1] C. Gravino, G. Scanniello, and G. Tortora, "Source-code comprehension tasks supported by UML design models:
Results from a controlled experiment and a differentiated replication,"
Journal of Visual Languages and Computing, vol. 28, pp. 23–38, 2015.
The experiments assess whether the comprehension of object oriented source-code increases used
with UML class and sequence diagrams produced in the software design phase.
If you use this data set please cite:
[1] C. Gravino, G. Scanniello, and G. Tortora, "Source-code comprehension tasks supported by UML
design models: Results from a controlled experiment and a differentiated replication,"
Journal of Visual Languages and Computing, vol. 28, pp. 23–38, 2015.
[2] Barbara Kitchenham, Lech Madeyski, Giuseppe Scanniello and Carmine Gravino, "The importance
of the Correlation between Results from Individual Participants in Crossover Experiments"
(to be submitted as of 2020).
}
\examples{
KitchenhamEtAl.CorrelationsAmongParticipants.Gravino15JVLC
}
\keyword{datasets}
