% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KitchenhamMadeyskiBrereton.ABBAReportedEffectSizes}
\alias{KitchenhamMadeyskiBrereton.ABBAReportedEffectSizes}
\title{KitchenhamMadeyskiBrereton.ABBAReportedEffectSizes data}
\format{
A text file with variables:
\describe{
\item{Study}{This field includes the study identifier of each of the 3 primary studies which were included in the systematic review. The studies are S3, S7 and S11.}
\item{Type}{This identifies the type of effect size used by the study authors. d  or g refer to dIG and gIG.}
\item{Source}{Always set to Rep. This identifies that the data was as reported by the primary study authors.}
\item{Design}{Mixed means different experiments in a particular family used different methods (onlyS3 used mixed methods and 4 experiments used the 4 group crossover and one used an independent groups design). ABBACO is the standard 2-group crossover design. }
\item{Exp1}{This is the reported standardised effect size for the first time period and the first experiment in the family.}
\item{Exp2}{This is the reported standardised effect size for the first time period and second experiment in the family.}
\item{Exp3}{This is the reported standardised effect size for the first time period and the third experiment in the family.}
\item{Exp4}{This is the reported standardised effect size for the first time period and the fourth experiment in the family. NA means there was no fourth experiment in the family.}
\item{Exp5}{This is the reported standardised effect size for the first time period and the fifth experiment in the family. NA means there was no fifth experiment in the family.}
}
}
\source{
\url{https://madeyski.e-informatyka.pl/reproducible-research/}
}
\usage{
KitchenhamMadeyskiBrereton.ABBAReportedEffectSizes
}
\description{
This data is used in the paper: Barbara Kitchenham, Lech Madeyski and Pearl Brereton. Meta-analysis for Families of Experiments: A Systematic Review and Reproducibility Assessment, Empirical Software Engineering (2019) doi:10.1007/s10664-019-09747-0.
This file holds the individual effect sizes for the first time period (or equivalently the first document), as reported by the 3 primary studies in the systematic review that reported results for each document/time period separately.
}
\examples{
KitchenhamMadeyskiBrereton.ABBAReportedEffectSizes

}
\keyword{datasets}
