% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KitchenhamMadeyskiBudgen16.DiffInDiffData}
\alias{KitchenhamMadeyskiBudgen16.DiffInDiffData}
\title{KitchenhamMadeyskiBudgen16.DiffInDiffData data}
\format{A data frame with variables:
\describe{
\item{Abstract}{The abstract identifier}
\item{Site}{A numeric identifier of the site}
\item{Treatment}{A three character alphanumeric identifying the jounral and time period of the abstract}
\item{Journal}{The journal in which the abstract was published: IST or JSS}
\item{Timeperiod}{The time period in which the abstract: 1 or 2}
\item{J1}{The identifer for the judge who made the next 2 assessments}
\item{J1Completeness}{The average completeness made by judge J1 based on the 8 completeness questions}
\item{J1Clarity}{The clarity assessment made by judge J1}
\item{J2}{The identifer for the judge who made the next 2 assessments}
\item{J2Completeness}{The average completeness made by judge J2 based on the 8 completeness questions}
\item{J2Clarity}{The clarity assessment made by judge J2}
\item{J3}{The identifer for the judge who made the next 2 assessments}
\item{J3Completeness}{The average completeness made by judge J3 based on the 8 completeness questions}
\item{J3Clarity}{The clarity assessment made by judge J3}
\item{J4}{The identifer for the judge who made the next 2 assessments}
\item{J4Completeness}{The average completeness made by judge J4 based on the 8 completeness questions}
\item{J4Clarity}{The clarity assessment made by judge J4}
\item{MeanCompleteness}{The mean of J1Completeness, J2Completeness, J3Completeness, J4Completeness}
\item{MedianCompleteness}{The median of J1Completeness, J2Completeness, J3Completeness, J4Completeness}
\item{MedianClarity}{The median clarity of J1Clarity, J2Clarity, J3Clarity, J4Clarity}
\item{MeanClarity}{The mean clarity of J1Clarity, J2Clarity, J3Clarity, J4Clarity}
\item{VarCompleteness}{The variance of J1Completeness, J2Completeness, J3Completeness, J4Completeness}
\item{VarClarity}{The variance clarity of J1Clarity, J2Clarity, J3Clarity, J4Clarity}
}}
\source{
\url{http://madeyski.e-informatyka.pl/reproducible-research/}
}
\usage{
KitchenhamMadeyskiBudgen16.DiffInDiffData
}
\description{
If you use this data set please cite this R package and the following paper when accepted: Barbara Kitchenham, Lech Madeyski, David Budgen, Jacky Keung, Pearl Brereton, Stuart Charters, Shirley Gibbs, and Amnart Pohthong, "Robust Statistical Methods for Empirical Software Engineering", Empirical Software Engineering (in press). DOI: 10.1007/s10664-016-9437-5 (http://dx.doi.org/10.1007/s10664-016-9437-5), URL: http://madeyski.e-informatyka.pl/download/KitchenhamMadeyskiESE.pdf
}
\details{
Data set was derived from the data reported in the SubjectData data set (subjectdata.txt). It contains the summary completeness and clarity data from 4 judges who assessed the same abstract. Only the initial 5 sites are included.

dinddata.txt
}
\examples{
KitchenhamMadeyskiBudgen16.DiffInDiffData

}
\keyword{datasets}

