% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KitchenhamMadeyskiBudgen16.PolishData}
\alias{KitchenhamMadeyskiBudgen16.PolishData}
\title{KitchenhamMadeyskiBudgen16.PolishData data}
\format{
A data frame with variables:
\describe{
\item{Abstract}{The abstract identifier}
\item{Site}{Numeric identifier for the site}
\item{Treatment}{The first three characters of the Abstract field which identifies the journal and time period of the abstract}
\item{Journal}{An acronym for the journal from which the abstract was obtained: IST or JSS}
\item{Timeperiod}{The Time period in which the abstract was found: 1 or 2}
\item{J1}{The identifier for the judge who made the next 2 assessments}
\item{J1Completeness}{The average completeness made by judge J1 based on the 8 completeness questions}
\item{J1Clarity}{The clarity assessment made by judge J1}
\item{J2}{The identifier for the judge who made the next 2 assessments}
\item{J2Completeness}{The average completeness made by judge J2 based on the 8 completeness questions}
\item{J2Clarity}{The clarity assessment made by judge J2}
\item{J3}{The identifier for the judge who made the next 2 assessments}
\item{J3Completeness}{The average completeness made by judge J3 based on the 8 completeness questions}
\item{J3Clarity}{The clarity assessment made by judge J3}
\item{J4}{The identifier for the judge who made the next 2 assessments}
\item{J4Completeness}{The average completeness made by judge J4 based on the 8 completeness questions}
\item{J4Clarity}{The clarity assessment made by judge J4}
\item{MedianCompleteness}{The median of J1Completeness, J2Completeness, J3Completeness, J4Completeness}
\item{MedianClarity}{The median of J1Clarity, J2Clarity, J3Clarity, J4Clarity}
}
}
\source{
\url{https://madeyski.e-informatyka.pl/reproducible-research/}
}
\usage{
KitchenhamMadeyskiBudgen16.PolishData
}
\description{
If you use this data set please cite this R package and the following paper when accepted: Barbara Kitchenham, Lech Madeyski, David Budgen, Jacky Keung, Pearl Brereton, Stuart Charters, Shirley Gibbs, and Amnart Pohthong, "Robust Statistical Methods for Empirical Software Engineering", Empirical Software Engineering, vol. 22, no.2, p. 579-630, 2017. DOI: 10.1007/s10664-016-9437-5 (https://dx.doi.org/10.1007/s10664-016-9437-5), URL: https://madeyski.e-informatyka.pl/download/KitchenhamMadeyskiESE.pdf
}
\details{
Data set derived from PolishSubjects data set collected at Wroclaw University. It summarizes the completeness and clarity data collected from 4 judges about the same abstract.

PolishData.txt
}
\examples{
KitchenhamMadeyskiBudgen16.PolishData

}
\keyword{datasets}
