% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MadeyskiKitchenham.MetaAnalysis.PBRvsCBRorAR}
\alias{MadeyskiKitchenham.MetaAnalysis.PBRvsCBRorAR}
\title{MadeyskiKitchenham.MetaAnalysis.PBRvsCBRorAR data form a set of primary studies on reading methods for software inspections. They were analysed by Lech Madeyski and Barbara Kitchenham, "How variations in experimental designs impact the construction of comparable effect sizes for meta-analysis", 2015.}
\format{
A data frame with 17 rows and 26 variables:
\describe{
\item{Study}{Name of empirical study}
\item{Ref.}{Reference to the paper reporting primary study or experimental run where data were originally reported}
\item{Teams}{The number of teams including both, PBR and Control teams}
\item{DesignDesc}{Experimental design description: Before-after, Between-groups, Cross-over}
\item{ExpDesign}{Experimental design: between-groups (BG), within-subjects cross-over (WSCO), within-subjects before-after (WSBA)}
\item{M_PBR}{The average proportion of defects found by teams using PBR}
\item{M_C}{The average proportion of defects found by teams using Control treatment: Check-Based Reading (CBR) or Ad-Hoc Reading (AR)}
\item{Diff}{The difference between M_PBR and M_C, i.e. Diff = M_PBR - M_C}
\item{Inc}{The percentage increase in defect rate detection, i.e. Inc=100*[(M_PBR-M_C)/M_C]}
\item{SD_C_ByAuthors}{The standard deviation of the control group values reported by the original Authors, i.e., obtained from the papers/raw data}
\item{SD_C}{The standard deviation of the control group values equals SD_C_ByAuthors for studies for which the data was available OR the weighted average of SD_C_ByAuthors (i.e., 0.169) for studies where SD_C_ByAuthors is missing.}
\item{V_C}{The variance of the Control group observations, i.e., the variance obtained from the teams using the Control method V_C=SD_C^2}
\item{V_D}{The variance of the unstandardized mean difference D (between the mean value for the treatment group and the mean value for the Control group)}
\item{SD_C_Alt}{This is the equivalent of SD_C (the standard deviation of the control group) based on a different variance for the student studies or the practitioner studies depending on the subject type of the study with the missing value.}
\item{V_Alt}{The variance of the mean difference in the meta-analysis based on SD_C_Alt}
\item{SS_C}{The sum of squares of the Control group values. For within subjects studies SS=V_C*(n-1). For between subjects studies SS=V_C*(n_C-1)}
\item{n_PBR}{The number of PBR teams}
\item{n_C}{The number of Control (CBR or AR) teams}
\item{ControlType}{Type of Control treatment: CRB or AR}
\item{ParticipantsType}{Type of participants: Engineers or Students}
\item{TeamType}{Type of team: Nominal or Real}
\item{TwoPersonTeamVsLargerTeam}{Reflects size of the teams: 2-PersonTeam or LargerTeam}
\item{ArtefactType}{The type of artefact: Requirements or Other}
\item{AssociatedWithBasili}{Whether study is associated with Basili (the forerunner): Yes or No}
\item{ControlType_Basili}{Combined ControlType and AssociatedWithBasili: AH_AssociatedWithBasili, CBR_AssociatedWithBasili, CBR_NotAssociatedWithBasili}
}
}
\source{
\url{https://madeyski.e-informatyka.pl/reproducible-research/}
}
\usage{
MadeyskiKitchenham.MetaAnalysis.PBRvsCBRorAR
}
\description{
If you use this data set please cite: Lech Madeyski and Barbara Kitchenham, "How variations in experimental designs impact the construction of comparable effect sizes for meta-analysis", 2015.
}
\examples{
MadeyskiKitchenham.MetaAnalysis.PBRvsCBRorAR

}
\keyword{datasets}
