% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{Function to pack the experiment}
\usage{
pack(pack_project = ".")
}
\arguments{
\item{pack_project}{Path to experiment directory. Defaults to '.'.}
}
\description{
This function allows you to package your experiment using packrat.
It then creates a Vagrantfile and starts a VM containing the experiment.
}
\examples{
pack()
}
\keyword{pack}
