% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parserDB-old.R, R/parserDB.R
\name{parserDB}
\alias{parserDB}
\title{Function to parse de json file into a relational provenance database}
\usage{
parserDB(con, prov)

parserDB(con, prov)
}
\arguments{
\item{con}{Connection to the provenance database}

\item{prov}{Path to the prov.json file}

\item{con}{Connection to the provenance database}

\item{prov}{Path to the prov.json file}
}
\description{
This contains the necessary function to read the json file containing
provenance information and insert this information into a relational database

This contains the necessary function to read the json file containing
provenance information and insert this information into a relational database
}
\keyword{creation,}
\keyword{database,}
\keyword{json}
\keyword{provenance,}
\keyword{sql,}
