% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPSimulation.R
\name{PHatonesidedTestStatistics}
\alias{PHatonesidedTestStatistics}
\title{PHatonesidedTestStatistics}
\usage{
PHatonesidedTestStatistics(
  effectsize,
  effectsize.variance,
  effectsize.df = 0,
  alpha = 0.05,
  alternative = "greater"
)
}
\arguments{
\item{effectsize}{The overall estimate of the centralized PHat (i.e. Phat-0.5) from a group of effect sizes to be meta-analysed}

\item{effectsize.variance}{The estimate of the variance of the overall estimate ofPHat}

\item{effectsize.df}{The total degrees of freedom for the set of effect sizes. If effectsize.df>0, the confidence intervals, pvalues and significance test use the t-distribution probability values. If effectsize.df=0 (default), the confidence intervals, the pvalues and significance test use the normal distribution probability values.}

\item{alpha}{The significance level (default 0.05) used to control the significance tests and calculation of confidence limits.}

\item{alternative}{Specifies the type of significance test and can take the values "less" or "greater" (default).}
}
\value{
ES.test The value of the t-statistic

ES.pvalue The p-value of the two-sided t-test if the parameter d.df>0, or the normal probability value if d.df=0

ES.sig The significance of the statistical test of the d.tvalue return value at the alpha level for one sided tests and aplha/2 for two sided tests as specified by the input parameter alternative.

ES.ci.lower The lower 100*(1-alpha/2)% confidence interval of the average centralized PHat based on the t-distribution probability values if effectsize.df>0 or normal probability values if effectsize.df=0

ES.ci.upper The upper 100*(1-alpha/2)% confidence interval of the average centralized PHat based on the t-distribution probability values if effectsize.df>0 or normal probability values if effectsize.df=0
}
\description{
This function is a helper function for meta-analysis of experiments using PHat as an effect size. It returns the 100*(1-alpha)% confidence intervals, the value of the t-test, the probability asssociated with the null hypothesis and the significance for a one-sided test. The direction of the one-sided test is determined by the parameter "alternative" which takes the values "greater" or "less".
}
\examples{
PHatES=mean(c(0.92,0.6,0.48,0.72,0.88))-0.5
PHatESvar=sum(c(0.01,0.04,0.05,0.04,0.01))/25
PHatdf=sum(c(6.63,6.63,5.08,5.61,8))
#PHatonesidedTestStatistics(effectsize=PHatES,effectsize.variance=PHatESvar,effectsize.df=PHatdf)
# A tibble: 1 x 5
#  ES.test ES.pvalue ES.sig ES.ci.lower ES.ci.upper
#    <dbl>     <dbl> <lgl>        <dbl>       <dbl>
#1    2.84   0.00389 TRUE        0.0888       0.351
#PHatonesidedTestStatistics(effectsize=PHatES,effectsize.variance=PHatESvar,effectsize.df=0,
# alternative="less")
# A tibble: 1 x 5
#  ES.test ES.pvalue ES.sig ES.ci.lower ES.ci.upper
#    <dbl>     <dbl> <lgl>        <dbl>       <dbl>
#1    2.84     0.998 FALSE       0.0926       0.347
}
\author{
Barbara Kitchenham and Lech Madeyski
}
