% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserve.R
\name{preserve}
\alias{preserve}
\title{Function to preserve the experiment}
\usage{
preserve(con = "~/prov.db",
  prov_json = "~/Dropbox/Artigos/2019/ER/experiment/modelr/prov_script/prov.json")
}
\arguments{
\item{con}{Connection to the porvenance database}

\item{prov_json}{Location of the prov.json file generated by the validate function}
}
\description{
This function allows you to preserve the experiment, detecting new installed applications
and creating a relational database to store provenance information
}
\examples{
preserve()
preserve(con = '~/prov.db', prov_json = '~/prov.json')
}
\keyword{preserve,}
\keyword{provenance}
