% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MadeyskiKitchenhamMetaAnalysis.R
\name{proportionOfSignificantTValuesUsingCorrectAnalysis}
\alias{proportionOfSignificantTValuesUsingCorrectAnalysis}
\title{proportionOfSignificantTValuesUsingCorrectAnalysis}
\usage{
proportionOfSignificantTValuesUsingCorrectAnalysis(data)
}
\arguments{
\item{data}{- data behind the plot returned by getSimulatedCrossoverDataSets() or stored in reproducer::KitchenhamMadeyski.SimulatedCrossoverDataSets}
}
\value{
plot described in description
}
\description{
Plots visualize the relationship between sample size, effect size and the proportion of significant t-values using the correct analysis. Function is used in a paper 'Effect Sizes and their Variance for AB/BA Crossover Design Studies' by Lech Madeyski and Barbara Kitchenham.
}
\examples{
data <- KitchenhamMadeyski.SimulatedCrossoverDataSets
myPlot <- proportionOfSignificantTValuesUsingCorrectAnalysis(data)
}
\author{
Barbara Kitchenham and Lech Madeyski
}
