% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MadeyskiKitchenhamMetaAnalysis.R
\name{proportionOfSignificantTValuesUsingIncorrectAnalysis}
\alias{proportionOfSignificantTValuesUsingIncorrectAnalysis}
\title{proportionOfSignificantTValuesUsingIncorrectAnalysis}
\usage{
proportionOfSignificantTValuesUsingIncorrectAnalysis(data)
}
\arguments{
\item{data}{- data behind the plot returned by getSimulatedCrossoverDataSets() or stored in reproducer::KitchenhamMadeyski.SimulatedCrossoverDataSets}
}
\value{
plot described in description
}
\description{
Plots visualize the relationship between sample size, effect size and the proportion of significant t-values using the incorrect analysis. Function is used in a paper "Effect Sizes and their Variance for AB/BA Crossover Design Studies" by Lech Madeyski and Barbara Kitchenham.
}
\examples{
data  <- KitchenhamMadeyski.SimulatedCrossoverDataSets
myPlot <- proportionOfSignificantTValuesUsingIncorrectAnalysis(data)
}
\author{
Barbara Kitchenham and Lech Madeyski
}

