% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Function to publish the experiment on zenodo}
\usage{
publish(token, title = "My publication title",
  description = "My description", uploadType = "publication",
  pubType = "", creatorFirstname = "Creator first name",
  creatorLastname = "Creator last name", license = "mit", bundle,
  prov_dir, prov_db = "~/prov.db")
}
\arguments{
\item{token}{User token (must be created in Zenodo).}

\item{title}{Title of the publication on Zenodo. It can be edited on the Zenodo platform after publication.}

\item{description}{Description of the publication on Zenodo. It can be edited on the Zenodo platform after publication.}

\item{uploadType}{Upload type on Zenodo. It can be edited on the Zenodo platform after publication.}

\item{pubType}{Publication type on Zenodo. It can be edited on the Zenodo platform after publication.}

\item{creatorFirstname}{Creator firstname. It can be edited on the Zenodo platform after publication.}

\item{creatorLastname}{Creator lastname It can be edited on the Zenodo platform after publication.}

\item{license}{Publication license. It can be edited on the Zenodo platform after publication.}

\item{bundle}{Path to compressed file containig the experiment (generated with packrat).}

\item{prov_dir}{Path to the provenance directory.}
}
\description{
This function allows you to publish your experiment in Zenodo.
It gathers the provenance files and database, the experiment bundled with packrat
and the Vagrant file with the system specifications to reproduce the experiment.
}
\examples{
publish()
}
\keyword{publish,}
\keyword{zenodo}
