% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducer.R
\name{readExcelSheet}
\alias{readExcelSheet}
\title{readExcelSheet}
\usage{
readExcelSheet(path, sheet, colNames)
}
\arguments{
\item{path}{Path to an Excel file, e.g. /User/lma/datasets/MyDataSet.xls}

\item{sheet}{Name of a sheet within an Excel file we want to read}

\item{colNames}{If TRUE, first row of data will be used as column names.}
}
\description{
Function reads data from an Excel file from a specified sheet
}
\examples{
myPath=system.file("extdata", "DataSet.xlsx", package = "reproducer")
Madeyski15SQJ.NDC<-readExcelSheet(path=myPath, sheet="Madeyski15SQJ.NDC", colNames=TRUE)
}
\author{
Lech Madeyski
}

