% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MadeyskiKitchenhamMetaAnalysis.R
\name{reproduceForestPlotRandomEffects}
\alias{reproduceForestPlotRandomEffects}
\title{reproduceForestPlotRandomEffects()}
\usage{
reproduceForestPlotRandomEffects()
}
\description{
Function reproduces Forest Plot of a Random-Effects Meta-analysis of Mean Differences.
}
\examples{
reproduceForestPlotRandomEffects()
}
\author{
Lech Madeyski
}
