% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MadeyskiKitchenhamMetaAnalysis.R
\name{reproduceMixedEffectsAnalysisWithExperimentalDesignModerator}
\alias{reproduceMixedEffectsAnalysisWithExperimentalDesignModerator}
\title{reproduceMixedEffectsAnalysisWithExperimentalDesignModerator()}
\usage{
reproduceMixedEffectsAnalysisWithExperimentalDesignModerator()
}
\description{
Function reproduces Mixed-Effects Analysis with Experimental Design as a Moderator.
}
\examples{
reproduceMixedEffectsAnalysisWithExperimentalDesignModerator()
}
\author{
Lech Madeyski
}
