% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MadeyskiKitchenhamMetaAnalysis.R
\name{reproduceMixedEffectsForestPlotWithExperimentalDesignModerator}
\alias{reproduceMixedEffectsForestPlotWithExperimentalDesignModerator}
\title{reproduceMixedEffectsForestPlotWithExperimentalDesignModerator()}
\usage{
reproduceMixedEffectsForestPlotWithExperimentalDesignModerator()
}
\description{
Function reproduces Forest Plot of a Mixed Effects Meta-analysis of Mean Differences with Experimental Design as a Moderator Variable.
}
\examples{
reproduceMixedEffectsForestPlotWithExperimentalDesignModerator()
}
\author{
Lech Madeyski
}

