% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MadeyskiKitchenhamMetaAnalysis.R
\name{reproduceSimulationResultsBasedOn500Reps1000Obs}
\alias{reproduceSimulationResultsBasedOn500Reps1000Obs}
\title{reproduceSimulationResultsBasedOn500Reps1000Obs}
\usage{
reproduceSimulationResultsBasedOn500Reps1000Obs()
}
\value{
data frame including the following simulation results:
# treatmentEffect.Ave - Average Technique Effect
# dRM.Ave - Average dRM
# dRM.Var - Variance of dRM
# dRM.Var.Ave - Average of var(dRM)
# dRM.Var.ModerateSampleSizeApprox -
# dIG.Ave - Average dIG
# dIG.Var - Variance of dIG
# dIG.Var.Ave - Average of var(dIG)
# dIG.Var.ModerateSampleSizeApprox -
# dIG.Var.CurtinAve -
}
\description{
Function to calculate simulation results based on 500 repetitions of 1000 observation samples. Function is used in a paper "Effect Sizes and their Variance for AB/BA Crossover Design Studies" by Lech Madeyski and Barbara Kitchenham.
}
\examples{
# return simulation results based on 500 repetitions of 1000 observation samples
simulationResultsTable500x1000<-reproduceSimulationResultsBasedOn500Reps1000Obs()
}
\author{
Lech Madeyski and Barbara Kitchenham
}

