% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MadeyskiKitchenhamMetaAnalysis.R
\name{reproduceTableWithEffectSizesBasedOnMeanDifferences}
\alias{reproduceTableWithEffectSizesBasedOnMeanDifferences}
\title{reproduceTableWithEffectSizesBasedOnMeanDifferences()}
\usage{
reproduceTableWithEffectSizesBasedOnMeanDifferences()
}
\description{
Function reproduces Table, which shows the effect sizes based on mean differences.
}
\examples{
reproduceTableWithEffectSizesBasedOnMeanDifferences()
}
\author{
Lech Madeyski
}

