% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MadeyskiKitchenhamMetaAnalysis.R
\name{reproduceTableWithPossibleModeratingFactors}
\alias{reproduceTableWithPossibleModeratingFactors}
\title{reproduceTableWithPossibleModeratingFactors()}
\usage{
reproduceTableWithPossibleModeratingFactors()
}
\description{
Function reproduces Table with possible moderating factors.
}
\examples{
reproduceTableWithPossibleModeratingFactors()
}
\author{
Lech Madeyski
}
