% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MadeyskiKitchenhamMetaAnalysis.R
\name{reproduceTableWithSourceDataByCiolkowski}
\alias{reproduceTableWithSourceDataByCiolkowski}
\title{reproduceTableWithSourceDataByCiolkowski}
\usage{
reproduceTableWithSourceDataByCiolkowski()
}
\description{
Function reproduces Table, which shows the effect sizes reported by Ciolkowski identifying the type of design used in each study.
}
\examples{
reproduceTableWithSourceDataByCiolkowski()
}
\author{
Lech Madeyski
}

