% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{searchForIndustryRelevantGitHubProjects}
\alias{searchForIndustryRelevantGitHubProjects}
\title{searchForIndustryRelevantGitHubProjects}
\usage{
searchForIndustryRelevantGitHubProjects(myToken)
}
\arguments{
\item{myToken}{A private token used to access GitHub}
}
\value{
selected GitHub projects
}
\description{
Function searches for industry relevant software projects available from GitHub. The function was used to deliver data set of software projects in an NCBiR project. More details are desribed in a report: Lech Madeyski, “Training data preparation method,” tech. rep., code quest (research project NCBiR POIR.01.01.01-00-0792/16), 2019, as well as a paper: Tomasz Lewowski and Lech Madeyski, "Creating evolving project data sets in software engineering", 2019.
If you use this function or the returned data set than please cite:
Tomasz Lewowski and Lech Madeyski, "Creating evolving project data sets in software engineering", 2019
}
\examples{
  #to run this function you need to use your own token as a parameter of the function
  #calculateSmallSampleSizeAdjustment("...") #use your own token as a parameter of the function
}
\author{
Lech Madeyski and Tomasz Lewowski
}
