% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{sqlFromFile}
\alias{sqlFromFile}
\title{Functions to create the provenance database}
\usage{
sqlFromFile(file)
}
\arguments{
\item{file}{File with the schema definition (sql file)}

\item{con}{Connection to the provenance database}
}
\description{
This contains the necessary functions to create the database file based on
the model decribed on db_schema.sql
}
\keyword{creation}
\keyword{database,}
\keyword{sql,}
