% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysisForFamiliesOfExperimentsSR.R
\name{transformHgtoR}
\alias{transformHgtoR}
\title{transformHgtoR}
\usage{
transformHgtoR(g, Nc, Nt)
}
\arguments{
\item{g}{A vector of Hegdes g values.}

\item{Nc}{A vector of numbers identifying the number of control condition participants in each group}

\item{Nt}{A vector of numbers identifying the number of treatment condition participants in each group}
}
\value{
value of point biserial r
}
\description{
The functions transforms a vector of Hedges g values to their equivalent point bi-serial values.
}
\examples{
transformHgtoR(0.4, 20, 20)
# [1] 0.1961161
}
\author{
Barbara Kitchenham and Lech Madeyski
}
