% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysisForFamiliesOfExperimentsSR.R
\name{transformHgtoZr}
\alias{transformHgtoZr}
\title{transformHgtoZr}
\usage{
transformHgtoZr(g, Nc, Nt)
}
\arguments{
\item{g}{value of Hedges' g}

\item{Nc}{the number of observations (participants) in the first (control) group}

\item{Nt}{the number of observations (participants) in the second (treatment) group}
}
\value{
value of normal approximation of point biserial r
}
\description{
The functions transforms a vector of Hedges g values to their normal approximation of point bi-serial values.
}
\examples{
transformHgtoZr(0.5, 20, 20)
# [1] 0.2474665
}
\author{
Barbara Kitchenham and Lech Madeyski
}
