% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysisForFamiliesOfExperimentsSR.R
\name{transformRtoHg}
\alias{transformRtoHg}
\title{transformRtoHg}
\usage{
transformRtoHg(r, Nc, Nt)
}
\arguments{
\item{r}{A vector of point bi-serial correlation values.}

\item{Nc}{A vector of the number of observations in the control condition for the related experiments.}

\item{Nt}{A vector of the number of observations in the treatment condition for the related experiments.}
}
\value{
value of Hedges' g
}
\description{
This function coverts a vector of point bi-serial r values with associated sample size information back to the mean difference effect size Hedges g.
}
\examples{
transformRtoHg(c(0.4, 0.2), c(20, 20), c(20, 20))
# [1] 0.8728716 0.4082483
}
\author{
Barbara Kitchenham and Lech Madeyski
}
