% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysisForFamiliesOfExperimentsSR.R
\name{transformRtoZr}
\alias{transformRtoZr}
\title{transformRtoZr}
\usage{
transformRtoZr(r)
}
\arguments{
\item{r}{A vector of r-values}
}
\value{
value of normal approximation of point biserial r
}
\description{
The function transforms a vector of point biserial r values to their normal approximation. It also works for the correlation r.
}
\examples{
transformRtoZr(0.4)
# [1] 0.4236489
Zr=transformRtoZr(c(0.4,0.2))
Zr
# [1] 0.4236489 0.2027326
}
\author{
Barbara Kitchenham and Lech Madeyski
}
