% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysisForFamiliesOfExperimentsSR.R
\name{transformZrtoHg}
\alias{transformZrtoHg}
\title{transformZrtoHg}
\usage{
transformZrtoHg(Zr, Nc, Nt)
}
\arguments{
\item{Zr}{the normal variate}

\item{Nc}{the number of observations (participants) in the first (control) group}

\item{Nt}{the number of observations (participants) in the second (treatment) group}
}
\value{
value of Hedges' g
}
\description{
Transforms Zr to Hedge's g.
}
\examples{
transformZrtoHg(0.5, 20, 20)
#[1] 1.042191
}
\author{
Barbara Kitchenham and Lech Madeyski
}
