% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysisForFamiliesOfExperimentsSR.R
\name{transformZrtoR}
\alias{transformZrtoR}
\title{transformZrtoR}
\usage{
transformZrtoR(zr)
}
\arguments{
\item{zr}{A vector of standard normal variates.}
}
\value{
value of point biserial r
}
\description{
The function transforms a vector of standardized normal variates to their equivalent r-values.
}
\examples{
transformZrtoR(0.4236489)
# [1] 0.4
transformZrtoR(c(0.4236489, 0.2027326))
# [1] 0.4 0.2
}
\author{
Barbara Kitchenham and Lech Madeyski
}
