% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Function to run the experiment and collect provenance}
\usage{
validate(script = "script.R", dir = ".")
}
\arguments{
\item{script}{Script to be executed. Defaults to 'script.R'}
}
\description{
This function allows you to run your experiment using rdt
to collect provenance information.
It then creates a directory containing the orignal script and the collected provenance data
}
\examples{
validate()
}
\keyword{provenance}
\keyword{validate,}
