% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Madeyski15EISEJ.StudProjects}
\alias{Madeyski15EISEJ.StudProjects}
\title{Madeyski15EISEJ.StudProjects data}
\format{A data frame with variables:
\describe{
\item{PROP}{The percentage of classes of proprietary (i.e., industrial) projects that must be tested in order to find 80\% of defects in case of software defect prediction models built on student (i.e., academic) projects.}
\item{NOTSTUD}{The percentage of classes of projects which are not student projects that must be tested in order to find 80\% of defects in case of software defect prediction models built on student (i.e., academic) projects.}
\item{STUD}{The percentage of classes of student (i.e., academic) projects that must be tested in order to find 80\% of defects in case of software defect prediction models built on student (i.e., academic) projects.}
\item{OPEN}{The percentage of classes of open source projects that must be tested in order to find 80\% of defects in case of software defect prediction models built on student (i.e., academic) projects.}
}}
\source{
\url{http://madeyski.e-informatyka.pl/reproducible-research/}
}
\usage{
Madeyski15EISEJ.StudProjects
}
\description{
If you use this data set please cite: Marian Jureczko and Lech Madeyski, "Cross-Project Defect Prediciton: An Empirical Study," (under review), 2015.
}
\examples{
Madeyski15EISEJ.StudProjects
}
\keyword{datasets}

