% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MadeyskiKitchenhamMetaAnalysis.R
\name{reproduceMixedEffectsAnalysisWithEstimatedVarianceAndExperimentalDesignModerator}
\alias{reproduceMixedEffectsAnalysisWithEstimatedVarianceAndExperimentalDesignModerator}
\title{reproduceMixedEffectsAnalysisWithEstimatedVarianceAndExperimentalDesignModerator()}
\usage{

  reproduceMixedEffectsAnalysisWithEstimatedVarianceAndExperimentalDesignModerator()
}
\description{
Function reproduces Mixed-Effects Analysis using Subject Specific Estimated Variance with Experimental Design as a Moderator.
}
\examples{
reproduceMixedEffectsAnalysisWithEstimatedVarianceAndExperimentalDesignModerator()
}
\author{
Lech Madeyski
}

