% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R, R/sqlFromFile.R
\name{sqlFromFile}
\alias{sqlFromFile}
\title{Functions to create the provenance database}
\usage{
sqlFromFile(file)

sqlFromFile(file)
}
\arguments{
\item{file}{File with the schema definition (sql file)}

\item{con}{Connection to the provenance database}

\item{con}{Connection to the provenance database}

\item{file}{File with the schema definition (sql file)}
}
\description{
This contains the necessary functions to create the database file based on
the model decribed on db_schema.sql

This contains the function that will read the sql file
with instructions to create the database and split into
various sql instructions. (DBI::dbSendQuery() does not accept multiple instructions)
}
\keyword{creation}
\keyword{database,}
\keyword{sql,}
